LayoutTests/storage/indexeddb/keyrange-required-arguments-expected.txt

Test IndexedDB keyrange required arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
PASS Boolean(indexedDB && IDBCursor && IDBDatabase && IDBDatabaseError && IDBDatabaseException && IDBFactory && IDBIndex && IDBKeyRange && IDBObjectStore && IDBRequest && IDBTransaction) is true

PASS IDBKeyRange.only(); threw exception TypeError: Not enough arguments.
PASS IDBKeyRange.lowerBound(); threw exception TypeError: Not enough arguments.
PASS IDBKeyRange.upperBound(); threw exception TypeError: Not enough arguments.
PASS IDBKeyRange.bound(1); threw exception TypeError: Not enough arguments.
PASS IDBKeyRange.bound(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
